Refactor app.css/ main so backgrounds are full on all views
