+++ src/App.css
@@ -20,6 +20,9 @@ main {
   align-items: center;
   display: flex;
   flex-direction: column;
+  align-items: center;
+  height: 100vh;
+  justify-content: center;
 }
 
 nav {
@@ -69,6 +72,7 @@ nav {
 
 .meadow-background {
   background-image: url('./assets/meadow.png');
+  background-size: cover;
 }
 
 @media screen and (max-width: 414px) and (max-height){
