+++ src/Components/Settings/Settings.css
@@ -1,10 +1,10 @@
 .theme-setting-container {
-    background-color: rgb(182, 180, 180);
-    border: 2px solid black;
-    height: 100px;
-    padding: 5px;
-    padding-top: 0px;
-  }
+  background-color: rgb(182, 180, 180);
+  border: 2px solid black;
+  height: 100px;
+  padding: 5px;
+  padding-top: 0px;
+}
 
 .settings-view {
   align-items: center;
