+++ src/index.css
@@ -2,5 +2,5 @@ body {
   margin: 0;
   background-position: top;
   margin: 0 auto;
-  background-repeat: no-repeat;
+  background-repeat: repeat;
 }
